+++ src/colloid_link_impl.h
@@ -4,7 +4,7 @@
  *
  *  Static inline function implementations. This file is included in
  *  colloid_link.h so that the relevant colloid_link_t definition is
- *  avaialble.
+ *  available.
  *
  *  Edinburgh Soft Matter and Statistical Physics Group and
  *  Edinburgh Parallel Computing Centre
@@ -31,10 +31,10 @@
  *
  *****************************************************************************/
 
-HDSI_ int colloid_links_array_i(const colloid_links_array_t * array, size_t index) {
+HDSI_ int colloid_links_array_i(const colloid_links_array_t * array, int index) {
 
   assert(0 <= index && index < array->max_links);
-  return array->i[index]; 
+  return array->i[index];
 }
 
 /*****************************************************************************
@@ -45,35 +45,35 @@ HDSI_ int colloid_links_array_i(const colloid_links_array_t * array, size_t inde
  *
  *****************************************************************************/
 
-HDSI_ int colloid_links_array_j(const colloid_links_array_t * array, size_t index) {
+HDSI_ int colloid_links_array_j(const colloid_links_array_t * array, int index) {
 
   assert(0 <= index && index < array->max_links);
   return array->j[index];
 }
 
 /***************************************************************************
- * 
+ *
  *  colloid_links_array__p
- * 
+ *
  *  Retrieves the p link at given index
- * 
+ *
  ***************************************************************************/
 
-HDSI_ int colloid_links_array_p(const colloid_links_array_t * array, size_t index) {
+HDSI_ int colloid_links_array_p(const colloid_links_array_t * array, int index) {
 
   assert(0 <= index && index < array->max_links);
   return array->p[index];
 }
 
 /***************************************************************************
- * 
+ *
  * colloid_links_array_status
- * 
+ *
  * Retrieves the link status at given index
- * 
+ *
  ***************************************************************************/
 
-HDSI_ int colloid_links_array_status(const colloid_links_array_t * array, size_t index) {
+HDSI_ int colloid_links_array_status(const colloid_links_array_t * array, int index) {
 
   assert(0 <= index && index < array->max_links);
 
@@ -81,17 +81,17 @@ HDSI_ int colloid_links_array_status(const colloid_links_array_t * array, size_t
 }
 
 /***************************************************************************
- * 
+ *
  * colloid_links_array_rb
- * 
+ *
  * Retrieves the rb link at given index
- * 
+ *
  ***************************************************************************/
 
-HDSI_ void colloid_links_array_rb(const colloid_links_array_t * array, size_t index, double *rb) {
+HDSI_ void colloid_links_array_rb(const colloid_links_array_t * array, int index, double * rb) {
 
   assert(0 <= index && index < array->max_links);
-  
+
   for (int i = 0; i < 3; i++) {
     rb[i] = array->rb[i][index];
   }
